check user for auth

+++ backend/controllers/goalController.js
@@ -1,14 +1,21 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // Get request (READ)
+// @desc    Get goals
+// @route   GET /api/goals
+// @access  Private
 const getGoals = asyncHandler(async (req, res) => {
 	const goals = await Goal.find({ user: req.user.id });
 	res.status(200).json(goals);
 });
 
 // Put request (CREATE)
+// @desc    Set goal
+// @route   POST /api/goals
+// @access  Private
 const setGoals = asyncHandler(async (req, res) => {
 	if (!req.body.text) {
 		res.status(400);
@@ -23,12 +30,29 @@ const setGoals = asyncHandler(async (req, res) => {
 });
 
 // Update request (UPDATE)
+// @desc    Update goal
+// @route   PUT /api/goals/:id
+// @access  Private
 const updateGoals = asyncHandler(async (req, res) => {
 	const goal = await Goal.findById(req.params.id);
 
 	if (!goal) {
 		res.status(400);
 		throw new Error("Goal not found!");
+	}
+
+	const user = await User.findById(req.user.id);
+
+	// Check for user
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+
+	// Make sure the logged in user matches the goal user
+	if (goal.user.toString() !== req.user.id) {
+		res.status(401);
+		throw new Error("User not authorized");
 	}
 
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -39,12 +63,29 @@ const updateGoals = asyncHandler(async (req, res) => {
 });
 
 // Delete request (DELETE)
+// @desc    Delete goal
+// @route   DELETE /api/goals/:id
+// @access  Private
 const deleteGoals = asyncHandler(async (req, res) => {
 	const goal = await Goal.findById(req.params.id);
 
 	if (!goal) {
 		res.status(400);
 		throw new Error("Goal not found!");
+	}
+
+	const user = await User.findById(req.user.id);
+
+	// Check for user
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+
+	// Make sure the logged in user matches the goal user
+	if (goal.user.toString() !== req.user.id) {
+		res.status(401);
+		throw new Error("User not authorized");
 	}
 
 	await goal.deleteOne();
